DEpopulate nipo.__init__ to make setup easier
