+++ nipo/attendance.py
@@ -7,7 +7,7 @@
 
 #The nipo calendar is here(https://calendar.google.com/calendar/b/1?cid=ODBmc2g3ajllMjg5amYybGJuYThmODk2dDBAZ3JvdXAuY2FsZW5kYXIuZ29vZ2xlLmNvbQ). It is a Google calendar where when you invite it to any events it is ging to accept given that this event doesnt clash with anything else on its calendar. This will be the default calendar that'll be used for timetabling in this project.
 
-from nipo import get_logger 
+from nipo.conf import get_logger 
 from nipo.db.schema import Student, Course, Module
 from datetime import datetime
 from numpy import uintc
