+++ nipo/conf.py
@@ -6,6 +6,37 @@ from nipo.bootstrap_conf import CONFIG_FILE_NAME
 CONFIG_FILENAME = CONFIG_FILE_NAME
 logger = get_logger("nipo_config")
 
+def get_configs(profile = 'DEFAULT'):
+	
+	curdir = os.path.dirname(__file__)	#The Path to this __init__ file. Trick from https://stackoverflow.com/questions/247770/how-to-retrieve-a-modules-path
+	curdir = os.path.abspath(curdir)
+	config = configparser.ConfigParser()
+	config.read(os.path.join(curdir,'..',CONFIG_FILENAME))
+	
+	return config[profile]
+
+def get_logger(loggerName):
+	log = logging.getLogger(loggerName)
+	# File handler which logs even debug messages
+	# To atttempt compression of old log files, try https://docs.python.org/3/howto/logging-cookbook.html#using-a-rotator-and-namer-to-customize-log-rotation-processing
+	config = get_configs()
+	log_location = os.path.join(config['Install Location'],'log','nipo.log')
+	fh = RotatingFileHandler(log_location, mode='a', maxBytes=1*1024*1024,backupCount=10, encoding=None, delay=0)
+	fh.setLevel(logging.DEBUG)
+	# Console handler that logs warnings or higher
+	ch = logging.StreamHandler()
+	ch.setLevel(logging.WARNING)
+	# create formatter and add it to the handlers
+	formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+	fh.setFormatter(formatter)
+	ch.setFormatter(formatter)
+	# add the handlers to the logger
+	log.addHandler(fh)
+	log.addHandler(ch)
+	log.setLevel(logging.DEBUG)
+	return log
+
+
 class NipoConfig:
 	'''
 	Class contains functionality for creating tables for a fresh install.
